fletuino/requirements.txt
'flet[all]'
paho-mqtt
pyserial
